+++ src/MotorVelocityControl.h
@@ -9,12 +9,12 @@
 #include "MotorLn298.h"
 
 namespace robopi{
-    constexpr float MAX_VEL_DF_DC = 16.755;//rad_s
+    constexpr double MAX_VEL_DF_DC = 16.755;//rad_s
 
 
     class VelocityEstimator {
         public:
-            virtual float estimate(double pos, double dT) = 0;
+            virtual double estimate(double pos, double dT) = 0;
     };
 
     class SlidingAverageFilter : public VelocityEstimator
@@ -34,7 +34,7 @@ namespace robopi{
             ~SlidingAverageFilter(){
                 delete _vs;
             }
-            float estimate(double pos, double dT) override;
+            double estimate(double pos, double dT) override;
         private:
             const double _size;
             double _posLast;
@@ -54,7 +54,7 @@ namespace robopi{
             _velIntegr(0.0f),
             _pos(0U){}
 
-            float estimate(double pos, double dT) override;
+            double estimate(double pos, double dT) override;
         private:
 
             double _pos;
@@ -68,7 +68,7 @@ namespace robopi{
     {
     public:
         MotorVelocityControl(MotorLn298* motor,Encoder* encoder, VelocityEstimator* velEstimator,
-            float kp, float ki, float kd,float maxVel = MAX_VEL_DF_DC):
+            double kp, double ki, double kd,double maxVel = MAX_VEL_DF_DC):
         _motor(motor),
         _velEstimator(velEstimator),
         _encoder(encoder),
@@ -88,19 +88,19 @@ namespace robopi{
          * Pass set point
          * @param velocity in [rad/s]
          */
-        void set(float velocity);
+        void set(double velocity);
 
         /**
          * Perform control loop iteration
          * @param dT time difference since last update [s]
          */
-        void update(float dT);
+        void update(double dT);
 
         /**
          * Get angular position from encoder
          * @return angle in [rad]
          */
-        float position() const {return _encoder->position();}
+        double position() const {return _encoder->position();}
 
         /**
          * Get position from encoder
@@ -113,13 +113,13 @@ namespace robopi{
        * Get filtered angular velocity
        * @return velocity in [rad/s]
        */
-        const float& velocity() const {return _velocityActual;}
+        const double& velocity() const {return _velocityActual;}
 
         /**
        * Get set point angular velocity
        * @return velocity in [rad/s]
        */
-        const float& velocitySet() const {return _velocitySet;}
+        const double& velocitySet() const {return _velocitySet;}
 
         /**
          * Stop motor
@@ -133,27 +133,27 @@ namespace robopi{
        * Get computed error
        * @return error in [rad/s]
        */
-        const float& error() { return _errLast;}
+        const double& error() { return _errLast;}
 
         /**
        * Get set point to pwm
        * @return pwm duty cycle in [%]
        */
-        float dutySet() { return _dutySet*100.0;}
+        double dutySet() { return _dutySet*100.0;}
 
         const MotorLn298* motor() const {return _motor;}
         const Encoder* encoder() const {return _encoder;}
         const VelocityEstimator* velocityEstimator() const {return _velEstimator;}
     protected:
 
-        float _kp,_ki,_kd;
-        float _errLast,_errIntegr;
+        double _kp,_ki,_kd;
+        double _errLast,_errIntegr;
 
         MotorLn298* _motor;
         Encoder* _encoder;
         VelocityEstimator* _velEstimator;
-        float _velocitySet, _dutySet;
-        float _velocityActual;
+        double _velocitySet, _dutySet;
+        double _velocityActual;
     
     };
